Move Table to output module and remove table module
